+++ api/Hmcr.Data/Database/Entities/HmrMimeType.cs
@@ -13,7 +13,7 @@ namespace Hmcr.Data.Database.Entities
         public decimal MimeTypeId { get; set; }
         public string MimeTypeCode { get; set; }
         public string Description { get; set; }
-        public decimal ConcurrencyControlNumber { get; set; }
+        public long ConcurrencyControlNumber { get; set; }
         public string DbAuditCreateUserid { get; set; }
         public DateTime DbAuditCreateTimestamp { get; set; }
         public string DbAuditLastUpdateUserid { get; set; }
